+++ tests/Kiota.Builder.Tests/Configuration/StructuredMimeTypesCollectionTests.cs
@@ -16,7 +16,7 @@ public sealed class StructuredMimeTypesCollectionTests
     public void ParsesWithOrWithoutPriorities()
     {
         var mimeTypes = new StructuredMimeTypesCollection(new[] { "application/json", "application/xml;q=0.8" });
-        Assert.Equal("application/json;q=1", mimeTypes.First(), StringComparer.OrdinalIgnoreCase);
+        Assert.Equal("application/json", mimeTypes.First(), StringComparer.OrdinalIgnoreCase);
         Assert.Equal("application/xml;q=0.8", mimeTypes.Last(), StringComparer.OrdinalIgnoreCase);
         Assert.DoesNotContain("application/atom+xml", mimeTypes);
         Assert.Equal(0.8f, mimeTypes.GetPriority("application/atom+xml"));
@@ -37,11 +37,13 @@ public sealed class StructuredMimeTypesCollectionTests
         Assert.Empty(mimeTypes);
     }
     [Theory]
-    [InlineData("application/json, application/xml, application/yaml", "application/json", "application/json;q=1")]
-    [InlineData("application/json, application/xml, application/yaml", "application/json,text/plain", "application/json;q=1")]
-    [InlineData("application/json, application/xml, application/yaml;q=0.8", "application/json,text/plain,application/yaml", "application/json;q=1,application/yaml;q=0.8")]
-    [InlineData("application/json, application/xml;q=0.9, application/yaml;q=0.8", "application/github+json", "application/github+json;q=1")]
-    [InlineData("application/vnd.topicus.keyhub+json;version=67, application/yaml;q=0.8", "application/vnd.topicus.keyhub+json;version=67", "application/vnd.topicus.keyhub+json;version=67;q=1")]
+    [InlineData("application/json, application/xml, application/yaml", "application/json", "application/json")]
+    [InlineData("application/json, application/xml, application/yaml", "application/json,text/plain", "application/json")]
+    [InlineData("application/json, application/xml, application/yaml;q=0.8", "application/json,text/plain,application/yaml", "application/json,application/yaml;q=0.8")]
+    [InlineData("application/json, application/xml;q=0.9, application/yaml;q=0.8", "application/github+json", "application/github+json")]
+    [InlineData("application/vnd.topicus.keyhub+json;version=67, application/yaml;q=0.8", "application/vnd.topicus.keyhub+json;version=67", "application/vnd.topicus.keyhub+json;version=67")]
+    [InlineData("application/vnd.topicus.keyhub+json, application/yaml;q=0.8", "application/vnd.topicus.keyhub+json;version=67", "application/vnd.topicus.keyhub+json;version=67")]
+    [InlineData("application/vnd.topicus.keyhub+json;version=67", "application/vnd.topicus.keyhub+json;version=67,application/vnd.topicus.keyhub+json;version=67,application/vnd.topicus.keyhub+xml;version=67", "application/vnd.topicus.keyhub+json;version=67")]
     public void MatchesAccept(string configuredTypes, string declaredTypes, string expectedTypes)
     {
         var mimeTypes = new StructuredMimeTypesCollection(configuredTypes.Split(',').Select(static x => x.Trim()));
@@ -49,6 +51,7 @@ public sealed class StructuredMimeTypesCollectionTests
         var deserializedExpectedTypes = expectedTypes.Split(',').Select(static x => x.Trim());
         foreach (var expectedType in deserializedExpectedTypes)
             Assert.Contains(expectedType, result);
+        Assert.Equal(result.Distinct(StringComparer.OrdinalIgnoreCase).Count(), result.Count());
     }
     [Theory]
     [InlineData("application/json, application/xml;q=0.9, application/yaml;q=0.8", "application/json", "application/json")]